# docs/swagger.yaml
openapi: 3.0.0
info:
  title: NTC Bus Tracking API
  version: 1.0.0
  description: Real-time GPS tracking system for NTC Sri Lanka inter-provincial buses with comprehensive bus operator management
  contact:
    name: NTC Sri Lanka
    url: https://www.ntc.gov.lk
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: http://api.route-sync.top/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the user
          example: "64f5a8b2c1d4e5f6a7b8c9d0"
        username:
          type: string
          description: Unique username
          example: "testadmin"
        email:
          type: string
          format: email
          description: User email address
          example: "admin@busroute.com"
        firstName:
          type: string
          description: User's first name
          example: "John"
        lastName:
          type: string
          description: User's last name
          example: "Doe"
        phone:
          type: string
          description: User's phone number
          example: "+94771234567"
        role:
          type: string
          enum: ["admin", "user", "bus_operator"]
          description: User role in the system
          example: "admin"
        isActive:
          type: boolean
          description: Whether the user account is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Account creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    Bus:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the bus
          example: "64f5a8b2c1d4e5f6a7b8c9d0"
        busId:
          type: string
          description: Auto-generated unique bus identifier
          example: "BUS123456"
          readOnly: true
        busNumber:
          type: string
          description: Bus registration number
          example: "NB-1234"
        routeId:
          oneOf:
            - type: string
              description: Route ID reference
              example: "64f5a8b2c1d4e5f6a7b8c9d1"
            - $ref: "#/components/schemas/Route"
          description: Associated route information
        operatorId:
          oneOf:
            - type: string
              description: Bus operator ID reference
              example: "64f5a8b2c1d4e5f6a7b8c9d2"
            - $ref: "#/components/schemas/User"
          description: Bus operator information
        capacity:
          type: number
          description: Total passenger capacity
          minimum: 1
          example: 45
        busType:
          type: string
          enum: ["Normal", "Semi Luxury", "Luxury", "Super Luxury"]
          description: Type of bus service
          example: "Semi Luxury"
        status:
          type: string
          enum: ["active", "inactive", "maintenance"]
          description: Current bus status
          example: "active"
        createdAt:
          type: string
          format: date-time
          description: Bus creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    Route:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the route
          example: "64f5a8b2c1d4e5f6a7b8c9d0"
        routeNumber:
          type: string
          description: Route identification number
          example: "100"
        origin:
          type: string
          description: Starting point of the route
          example: "Colombo"
        destination:
          type: string
          description: End point of the route
          example: "Kandy"
        distance:
          type: number
          description: Total route distance in kilometers
          minimum: 1
          example: 115
        estimatedDuration:
          type: number
          description: Estimated travel time in minutes
          minimum: 1
          example: 180
        isActive:
          type: boolean
          description: Whether the route is currently active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Route creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
    Location:
      type: object
      properties:
        _id:
          type: string
          description: Unique identifier for the location record
          example: "64f5a8b2c1d4e5f6a7b8c9d0"
        busId:
          oneOf:
            - type: string
              description: Bus ID reference
              example: "64f5a8b2c1d4e5f6a7b8c9d1"
            - $ref: "#/components/schemas/Bus"
          description: Associated bus information
        location:
          type: object
          description: GeoJSON Point format location
          properties:
            type:
              type: string
              enum: ["Point"]
              example: "Point"
            coordinates:
              type: array
              description: Longitude and latitude coordinates
              items:
                type: number
              minItems: 2
              maxItems: 2
              example: [79.8612, 6.9271]
        speed:
          type: number
          description: Current speed in km/h
          minimum: 0
          example: 45
        heading:
          type: number
          description: Direction in degrees (0-360)
          minimum: 0
          maximum: 360
          example: 180
        timestamp:
          type: string
          format: date-time
          description: Location record timestamp
        createdAt:
          type: string
          format: date-time
          description: Record creation timestamp
...
    LocationUpdateRequest:
      type: object
      required:
        - longitude
        - latitude
      properties:
        longitude:
          type: number
          description: Longitude coordinate
          minimum: -180
          maximum: 180
          example: 79.8612
        latitude:
          type: number
          description: Latitude coordinate
          minimum: -90
          maximum: 90
          example: 6.9271
        speed:
          type: number
          description: Current speed in km/h
          minimum: 0
          maximum: 200
          example: 45
        heading:
          type: number
          description: Direction in degrees
          minimum: 0
          maximum: 360
          example: 180
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: true
        message:
          type: string
          description: Success message
          example: "Operation completed successfully"
        data:
          type: object
          description: Response data
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Operation success status
          example: false
        message:
          type: string
          description: Error message
          example: "Error occurred"
        errors:
          type: array
          description: Detailed error information
          items:
            type: object
            properties:
              field:
                type: string
                example: "email"
              message:
                type: string
                example: "Email is required"
...
paths:
...
  /locations/bus/{busId}/date/{date}:
    get:
      tags:
        - Locations
      summary: Get bus location tracking for specific date
      description: Retrieve all location records for a bus on a specific date
      security:
        - bearerAuth: []
      parameters:
        - name: busId
          in: path
          required: true
          description: Bus ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - name: date
          in: path
          required: true
          description: Date in YYYY-MM-DD format
          schema:
            type: string
            format: date
            pattern: '^\d{4}-\d{2}-\d{2}$'
          example: "2025-09-29"
      responses:
        "200":
          description: Location tracking data retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      count:
                        type: number
                        description: Number of location records
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Location"
  /locations/bus/{busId}/update:
    post:
      tags:
        - Locations
      summary: Update bus location (POST)
      description: Create new location record for a bus
      security:
        - bearerAuth: []
      parameters:
        - name: busId
          in: path
          required: true
          description: Bus ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdateRequest"
            examples:
              location_update:
                summary: Standard location update
                value:
                  longitude: 79.8612
                  latitude: 6.9271
                  speed: 45
                  heading: 180
      responses:
        "201":
          description: Location updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Location"

    put:
      tags:
        - Locations
      summary: Update bus location (PUT)
      description: Update bus location using PUT method
      security:
        - bearerAuth: []
      parameters:
        - name: busId
          in: path
          required: true
          description: Bus ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdateRequest"
      responses:
        "200":
          description: Location updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /locations/bus/{busId}/latest:
    get:
      tags:
        - Locations
      summary: Get latest location for a bus
      description: Retrieve the most recent location record for a specific bus
      security:
        - bearerAuth: []
      parameters:
        - name: busId
          in: path
          required: true
          description: Bus ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
      responses:
        "200":
          description: Latest location retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Location"
        "404":
          description: No location data found for this bus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /locations/bus/{busId}/history:
    get:
      tags:
        - Locations
      summary: Get location history for a bus
      description: Retrieve historical location data for a bus with optional date filtering
      security:
        - bearerAuth: []
      parameters:
        - name: busId
          in: path
          required: true
          description: Bus ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - name: startDate
          in: query
          description: Start date for filtering
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date for filtering
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Maximum number of records
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        "200":
          description: Location history retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      count:
                        type: number
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Location"
  /locations/buses/active:
    get:
      tags:
        - Locations
      summary: Get all active bus locations
      description: Retrieve latest location data for all active buses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active bus locations retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      count:
                        type: number
                        description: Number of active buses
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Location"
  /locations/bus/{busId}/stats:
    get:
      tags:
        - Locations
      summary: Get location statistics for a bus
      description: Retrieve statistical information about a bus's location data
      security:
        - bearerAuth: []
      parameters:
        - name: busId
          in: path
          required: true
          description: Bus ID
          schema:
            type: string
            pattern: "^[0-9a-fA-F]{24}$"
        - name: startDate
          in: query
          description: Start date for statistics
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          description: End date for statistics
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Location statistics retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          totalRecords:
                            type: number
                            description: Total number of location records
                          averageSpeed:
                            type: number
                            description: Average speed in km/h
                          maxSpeed:
                            type: number
                            description: Maximum speed recorded
                          totalDistance:
                            type: number
                            description: Total distance traveled in km
                          firstRecord:
                            type: string
                            format: date-time
                            description: Timestamp of first record
                          lastRecord:
                            type: string
                            format: date-time
                            description: Timestamp of last record
...